adding status change propery for bulk delete

+++ mvc/views/marketzone/index.php
@@ -821,18 +821,22 @@ $('#mytree').jstree({
             });
 
 
-
 $("#bulkDelete").on('click',function() { // bulk checked
         var status = this.checked;
         $(".deleteRow").each( function() {
           if(status == 1 && $(this).prop('checked')) {
                 
           } else {
-            $(this).prop("checked",status);
-            $(this).not("#bulkDelete").closest('tr').toggleClass('rowselected');
+                if (status == false && $(this).prop('checked') == false) {
+
+                } else {
+                         $(this).prop("checked",status);
+                        $(this).not("#bulkDelete").closest('tr').toggleClass('rowselected');
+                }
          }
         });
     });
+
 
 
     $('#deleteTriger').on("click", function(event){ // triggering delete one by one
